chore: remove meteor.startup from unfollow message action
