+++ apps/meteor/client/components/message/toolbar/MessageToolbar.tsx
@@ -17,6 +17,7 @@ import { usePinMessageAction } from './usePinMessageAction';
 import { useStarMessageAction } from './useStarMessageAction';
 import { useUnstarMessageAction } from './useUnstarMessageAction';
 import { useWebDAVMessageAction } from './useWebDAVMessageAction';
+import { useUnFollowMessageAction } from '../../../../app/threads/client/messageAction/useUnFollowMessageAction';
 import type { MessageActionContext } from '../../../../app/ui-utils/client/lib/MessageAction';
 import { MessageAction } from '../../../../app/ui-utils/client/lib/MessageAction';
 import { useEmojiPickerData } from '../../../contexts/EmojiPickerContext';
@@ -98,6 +99,7 @@ const MessageToolbar = ({
 	useStarMessageAction(message, { room, user });
 	useUnstarMessageAction(message, { room, user });
 	usePermalinkStar(message, { subscription, user });
+	useUnFollowMessageAction(message, { room, user, context });
 
 	useJumpToMessageContextAction(message, {
 		id: 'jump-to-message',
